Move logic to pkg/logs
